requirements.txt
google-auth
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
kafka-python
pyspark==3.4.0
apache-airflow-providers-apache-spark==4.1.0